feat(editor): show C/I/A on asset hover in selected-component sidebar
